moving general fx to utils
